requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.20
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
orjson==3.11.7
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.0.1
python-multipart==0.0.22
python-telegram-bot==21.6
PyYAML==6.0.3
rich==14.3.2
rich-toolkit==0.18.1
shellingham==1.5.4
SQLAlchemy==2.0.30
starlette==0.37.2
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
uvicorn==0.30.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
